Extract top-block upsert into yaml/top_block.py
Fifth PR in the helpers/yaml split arc (stacked on #803).
Moves _locate_top_block, _find_prepend_anchor, and
upsert_yaml_leaf_under_top_block byte-for-byte from
yaml/__init__.py into the new yaml/top_block.py submodule.

top_block.py imports the scalar helpers it needs directly from
.scalar (ESPHOME_YAML_INDENT, YamlUpsertNotSupportedError,
_safe_yaml_scalar, read_yaml_scalar) and the substitution
fallback from .substitution (rewrite_name_or_substitution).

Removes a stale orphaned comment block (the
ESPHOME_YAML_INDENT mirror-frontend note) that scalar.py's
#803 carried but __init__.py's copy was orphaned above the
top_block functions.

Re-exports upsert_yaml_leaf_under_top_block at the bottom of
__init__.py in PEP 484 redundant-alias form so external
callers (devices/mutations_clone.py, tests) keep using
"from .helpers.yaml import upsert_yaml_leaf_under_top_block"
unchanged.

__init__.py: 683 -> 557 lines (-126). top_block.py: 136 lines.

All 3404 non-e2e tests pass; ruff + mypy clean.
